Added ReviewService, Added All Review functions to Main Controller, Added Review JSPs

+++ src/main/webapp/WEB-INF/gamesView.jsp
@@ -15,6 +15,7 @@
 	<link rel="stylesheet" type="text/css" href="/css/style.css" />
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
 	<link rel="stylesheet" href="https://cdn.datatables.net/1.13.6/css/jquery.dataTables.css" />
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.css"/>
 	<script src="https://code.jquery.com/jquery-3.7.0.js"></script>
 </head>
 <body>
@@ -74,20 +75,26 @@
         </div>
     </nav>
 	<div class="container-fluid d-flex flex-column justify-content-center mt-5">
-		<img src="${game.getImgUrl()}" style="width:100%; height:auto;">
+		<img src="${game.getImgUrl()}" style="width:100%; height:auto;" class="shadow-lg">
 		<h3 class="my-3 text-center">${game.getTitle()}</h3>
 		<c:if test="${currentUser.getId() == game.getUser().getId() }">
 			<div class="text-center mb-3">
 				<a href="/games/edit/${game.getId()}"><button class="btn btn-sm btn-warning">Edit</button></a>
-				<a href="/games/delete/${game.getId()}"><button class="btn btn-sm btn-danger">Delete</button></a>
+				<a href="/games/delete/${game.getId()}"><button class="btn btn-sm btn-danger" onclick="return confirm('Are you sure you want to delete this game?')">Delete</button></a>
 			</div>
 		</c:if>
 		<div class="d-flex align-items-center justify-content-between mb-3">
 			<a href="/"><button class="btn btn-sm btn-warning">Add to Owned Games</button></a>
 			<div class="d-flex">
 				<p>${game.getLikers().size()} people like this game.</p>
-				<a href="/games/like/${game.getId()}"><button class="btn btn-sm btn-primary ms-3">Like</button></a>
-				<a href="/games/unlike/${game.getId()}"><button class="btn btn-sm btn-primary ms-3">Unlike</button></a>
+				<c:choose>
+					<c:when test="${!game.getLikers().contains(currentUser)}">
+						<a href="/games/like/${game.getId()}"><button class="btn btn-sm btn-primary ms-3">Like</button></a>
+					</c:when>
+					<c:otherwise>
+						<a href="/games/unlike/${game.getId()}"><button class="btn btn-sm btn-primary ms-3">Unlike</button></a>
+					</c:otherwise>
+				</c:choose>
 			</div>
 		</div>
 	</div>
@@ -105,27 +112,95 @@
 			</tbody>
 		</table>
 		<div>
-			<a href=""><button class="btn btn-success">Write a Review</button></a>
+			<a href="/games/${game.getId()}/review"><button class="btn btn-success">Write a Review</button></a>
 		</div>
 		<div class="mt-5">
-			<h4 class="mb-5">All User Reviews:</h4>
-			<h4>Average Rating:</h4>	
+			<h4 class="mb-3">All User Reviews:</h4>
+			<c:choose>
+				<c:when test="${average >= 1.0 && average <= 1.49}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 1.5 && average <= 1.99}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star-half"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 2.0 && average <= 2.49}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 2.5 && average <= 2.99}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star-half"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 3.0 && average <= 3.49}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 3.5 && average <= 3.99}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star-half"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 4.0 && average <= 4.49}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average >= 4.5 && average <= 4.99}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star-half"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:when test="${average == 5}">
+					<h5>Average Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span> ${average} out of 5.</h5>
+				</c:when>
+				<c:otherwise>
+				</c:otherwise>
+			</c:choose>
 		</div>
-		<div>
+		<div class="mb-5">
 			<c:if test="${empty game.getReviews()}">
-			<p>This game has no reviews yet.</p>
+				<p>This game has no reviews yet.</p>
 			</c:if>
-			<table class="mb-4 table table-hover fs-5 align-middle border rounded overflow-hidden bg-gradient shadow">
-				<tbody>
-					<tr>
-						<td>Username</td>
-						<td>
-							<p>Rating:</p>
-							<p>Game Review</p>
-						</td>
-					</tr>
-				</tbody>
-			</table>
+			<c:if test="${not empty game.getReviews()}">
+				<table class="mb-4 table table-hover fs-5 align-middle border rounded overflow-hidden bg-gradient shadow">
+					<tbody>
+						<c:forEach var="oneReview" items="${game.getReviews()}">
+						<tr>
+							<td class="mt-3">
+								<div class="d-flex flex-column justify-content-center align-items-center">
+									<img src="${oneReview.getUser().getImgUrl()}" class="profile-pic"/>
+									<p>${oneReview.getUser().getUsername()}</p>
+								</div>
+							</td>
+							<td>
+								<c:choose>
+									<c:when test="${oneReview.getRating() == 1}">
+										<p>Rating: <span class="star"><i class="fa-solid fa-star"></i></span></p>
+									</c:when>
+									<c:when test="${oneReview.getRating() == 2}">
+										<p>Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span></p>
+									</c:when>
+									<c:when test="${oneReview.getRating() == 3}">
+										<p>Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span></p>
+									</c:when>
+									<c:when test="${oneReview.getRating() == 4}">
+										<p>Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span></p>
+									</c:when>
+									<c:when test="${oneReview.getRating() == 5}">
+										<p>Rating: <span class="star"><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i><i class="fa-solid fa-star"></i></span></p>
+									</c:when>
+								</c:choose>
+								<p>Posted: <fmt:formatDate value="${oneReview.getCreatedAt()}" pattern="MMMM dd, yyyy" /></p>
+								<p>${oneReview.getGameReview()}</p>
+							</td>
+							<c:if test="${currentUser.getId() == oneReview.getUser().getId()}">
+								<td>
+									<div>
+										<a href="/games/${game.getId()}/review/edit/${oneReview.getId()}"><button class="btn btn-sm btn-warning my-2">Edit Review</button></a>
+										<a href="/games/${game.getId()}/review/delete/${oneReview.getId()}"><button class="btn btn-sm btn-danger" onclick="return confirm('Are you sure you want to delete this review?')">Delete Review</button></a>
+									</div>
+								</td>
+							</c:if>
+							<c:if test="${currentUser.getId() != oneReview.getUser().getId()}">
+								<td>
+								</td>
+							</c:if>
+						</tr>
+						</c:forEach>
+					</tbody>
+				</table>
+			</c:if>
 		</div>
 	</div>
 	<script>
